[RT] Object -> Retainable, add refernece counter

+++ apps/dgemm/kernel.cl
@@ -1,11 +1,11 @@
 #pragma OPENCL EXTENSION cl_khr_fp64 : enable
 __kernel void ijk(__global double* restrict C, __global double* restrict A, __global double* restrict B) {
-  int i = get_global_id(1);
-  int j = get_global_id(0);
-  int SIZE = get_global_size(0);
+  size_t i = get_global_id(1);
+  size_t j = get_global_id(0);
+  size_t SIZE = get_global_size(0);
 
   double sum = 0.0;
-  for (int k = 0; k < SIZE; k++) {
+  for (size_t k = 0; k < SIZE; k++) {
     sum += A[i * SIZE + k] * B[k * SIZE + j];
   }
   C[i * SIZE + j] = sum;
